chore(apiserver): fetcher client rename
